rename authentication-stategies into authentication
