+++ server/src/authentication/jwt-strategy.ts
